+++ migrations/20220621140353_email_frequency_setting.down.sql
@@ -1,0 +1,4 @@
+DELETE FROM
+    user_setting
+WHERE
+    setting = 'email-frequency';