+++ queries/pending_notification/create.sql
@@ -1,0 +1,4 @@
+INSERT INTO
+    pending_notification (owner_id, user_event_id)
+VALUES
+    ($1, $2) RETURNING id;