+++ queries/pending_notification/remove.sql
@@ -1,0 +1,4 @@
+DELETE FROM
+    pending_notification
+WHERE
+    id = any($1);