+++ queries/user_event/resolve.sql
@@ -1,0 +1,6 @@
+SELECT
+    *
+FROM
+    user_event
+WHERE
+    id = any($1);