+++ content/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.11.0 — June 21st, 2022
+
+* Added  daily or weekly digest feature to reduce frequency of emailed alerts.
+* Fixed an issue where the FAQ link would only be visible to authenticated
+  users.
+
 ## v0.10.2 — June 20th, 2022
 
 * Added a frequently asked questions page.